Fan out statespace core tests to mirror the module split
Shared fixtures move to a conftest so the three test modules can be read independently; make_statespace_mod moves to test_utilities alongside the other cross-module helpers.
